Support ASP.Net Core
Currently `jarPath` is passed into `javaRunArgs` via `ResolveJarFiles`. `jarPath` is a relative path, and cannot be resolved on asp.net core. In this change `ResolveJarFiles` has been updated to use the full path for the required jar file, relative to the execution directory. Then resolves the options-specified jar file as is, or resolves it as a full path relative to the execution directory.